requirements.txt
flask
numpy
gunicorn
tensorflow-cpu